+++ src/Faithlife.Data/DbValueTypeStrategy.cs
@@ -10,5 +10,6 @@ namespace Faithlife.Data
 		Dynamic,
 		Dictionary,
 		Stream,
+		PositionalRecord,
 	}
 }