+++ tests/Faithlife.Data.Tests/DataRecordExtensionsTests.cs
@@ -6,6 +6,7 @@ using System.Data.SQLite;
 using System.IO;
 using System.Linq;
 using FluentAssertions;
+using FluentAssertions.Equivalency;
 using NUnit.Framework;
 using static FluentAssertions.FluentActions;
 
@@ -25,7 +26,7 @@ namespace Faithlife.Data.Tests
 			// get non-nulls
 			reader.Read().Should().BeTrue();
 
-			reader.Get<string>(0, 1).Should().Be(s_record.TheString);
+			reader.Get<string>(0, 1).Should().Be(s_dto.TheString);
 
 			// get nulls
 			reader.Read().Should().BeTrue();
@@ -44,11 +45,11 @@ namespace Faithlife.Data.Tests
 			// get non-nulls
 			reader.Read().Should().BeTrue();
 
-			reader.Get<int>(1, 1).Should().Be(s_record.TheInt32);
-			reader.Get<long>(2, 1).Should().Be(s_record.TheInt64);
-			reader.Get<bool>(3, 1).Should().Be(s_record.TheBool);
-			reader.Get<float>(4, 1).Should().Be(s_record.TheSingle);
-			reader.Get<double>(5, 1).Should().Be(s_record.TheDouble);
+			reader.Get<int>(1, 1).Should().Be(s_dto.TheInt32);
+			reader.Get<long>(2, 1).Should().Be(s_dto.TheInt64);
+			reader.Get<bool>(3, 1).Should().Be(s_dto.TheBool);
+			reader.Get<float>(4, 1).Should().Be(s_dto.TheSingle);
+			reader.Get<double>(5, 1).Should().Be(s_dto.TheDouble);
 
 			// get nulls
 			reader.Read().Should().BeTrue();
@@ -71,11 +72,11 @@ namespace Faithlife.Data.Tests
 			// get non-nulls
 			reader.Read().Should().BeTrue();
 
-			reader.Get<int?>(1, 1).Should().Be(s_record.TheInt32);
-			reader.Get<long?>(2, 1).Should().Be(s_record.TheInt64);
-			reader.Get<bool?>(3, 1).Should().Be(s_record.TheBool);
-			reader.Get<float?>(4, 1).Should().Be(s_record.TheSingle);
-			reader.Get<double?>(5, 1).Should().Be(s_record.TheDouble);
+			reader.Get<int?>(1, 1).Should().Be(s_dto.TheInt32);
+			reader.Get<long?>(2, 1).Should().Be(s_dto.TheInt64);
+			reader.Get<bool?>(3, 1).Should().Be(s_dto.TheBool);
+			reader.Get<float?>(4, 1).Should().Be(s_dto.TheSingle);
+			reader.Get<double?>(5, 1).Should().Be(s_dto.TheDouble);
 
 			// get nulls
 			reader.Read().Should().BeTrue();
@@ -122,10 +123,10 @@ namespace Faithlife.Data.Tests
 
 			reader.Read().Should().BeTrue();
 
-			Invoking(() => reader.Get<ItemRecord>(-1, 2)).Should().Throw<ArgumentException>();
-			Invoking(() => reader.Get<ItemRecord>(2, -1)).Should().Throw<ArgumentException>();
-			Invoking(() => reader.Get<ItemRecord>(7, 1)).Should().Throw<ArgumentException>();
-			Invoking(() => reader.Get<ItemRecord>(8, 0)).Should().Throw<ArgumentException>();
+			Invoking(() => reader.Get<ItemDto>(-1, 2)).Should().Throw<ArgumentException>();
+			Invoking(() => reader.Get<ItemDto>(2, -1)).Should().Throw<ArgumentException>();
+			Invoking(() => reader.Get<ItemDto>(7, 1)).Should().Throw<ArgumentException>();
+			Invoking(() => reader.Get<ItemDto>(8, 0)).Should().Throw<ArgumentException>();
 		}
 
 		[Test]
@@ -153,7 +154,7 @@ namespace Faithlife.Data.Tests
 			reader.Read().Should().BeTrue();
 
 			Invoking(() => reader.Get<(string, int)>(0, 1)).Should().Throw<InvalidOperationException>();
-			reader.Get<(string?, int)>(0, 2).Should().Be((s_record.TheString, s_record.TheInt32));
+			reader.Get<(string?, int)>(0, 2).Should().Be((s_dto.TheString, s_dto.TheInt32));
 			Invoking(() => reader.Get<(string, int)>(0, 3)).Should().Throw<InvalidOperationException>();
 		}
 
@@ -168,7 +169,7 @@ namespace Faithlife.Data.Tests
 			// get non-nulls
 			reader.Read().Should().BeTrue();
 
-			reader.Get<byte[]>(6, 1).Should().Equal(s_record.TheBlob);
+			reader.Get<byte[]>(6, 1).Should().Equal(s_dto.TheBlob);
 
 			// get nulls
 			reader.Read().Should().BeTrue();
@@ -189,8 +190,8 @@ namespace Faithlife.Data.Tests
 
 			var bytes = new byte[100];
 			using (var stream = reader.Get<Stream>(6, 1))
-				stream.Read(bytes, 0, bytes.Length).Should().Be(s_record.TheBlob!.Length);
-			bytes.Take(s_record.TheBlob!.Length).Should().Equal(s_record.TheBlob);
+				stream.Read(bytes, 0, bytes.Length).Should().Be(s_dto.TheBlob!.Length);
+			bytes.Take(s_dto.TheBlob!.Length).Should().Equal(s_dto.TheBlob);
 
 			// get nulls
 			reader.Read().Should().BeTrue();
@@ -210,9 +211,9 @@ namespace Faithlife.Data.Tests
 			reader.Read().Should().BeTrue();
 
 			reader.Get<(string?, int, long, bool, float, double)>(0, 6)
-				.Should().Be((s_record.TheString, s_record.TheInt32, s_record.TheInt64, s_record.TheBool, s_record.TheSingle, s_record.TheDouble));
+				.Should().Be((s_dto.TheString, s_dto.TheInt32, s_dto.TheInt64, s_dto.TheBool, s_dto.TheSingle, s_dto.TheDouble));
 			reader.Get<(string?, int, long, bool, float, double)>(..^1)
-				.Should().Be((s_record.TheString, s_record.TheInt32, s_record.TheInt64, s_record.TheBool, s_record.TheSingle, s_record.TheDouble));
+				.Should().Be((s_dto.TheString, s_dto.TheInt32, s_dto.TheInt64, s_dto.TheBool, s_dto.TheSingle, s_dto.TheDouble));
 
 			// get nulls
 			reader.Read().Should().BeTrue();
@@ -235,25 +236,25 @@ namespace Faithlife.Data.Tests
 			reader.Read().Should().BeTrue();
 
 			// DTO
-			reader.Get<ItemRecord>(0, 7).Should().BeEquivalentTo(s_record);
-			reader.Get<ItemRecord>(0, 1).Should().BeEquivalentTo(new ItemRecord { TheString = s_record.TheString });
-			reader.Get<ItemRecord>(0, 0).Should().BeNull();
-			reader.Get<ItemRecord>(7, 0).Should().BeNull();
+			reader.Get<ItemDto>(0, 7).Should().BeEquivalentTo(s_dto);
+			reader.Get<ItemDto>(0, 1).Should().BeEquivalentTo(new ItemDto { TheString = s_dto.TheString });
+			reader.Get<ItemDto>(0, 0).Should().BeNull();
+			reader.Get<ItemDto>(7, 0).Should().BeNull();
 
 			// tuple with DTO
-			var tuple = reader.Get<(string, ItemRecord, bool)>(0, 4);
-			tuple.Item1.Should().Be(s_record.TheString);
-			tuple.Item2.Should().BeEquivalentTo(new ItemRecord { TheInt32 = s_record.TheInt32, TheInt64 = s_record.TheInt64 });
+			var tuple = reader.Get<(string, ItemDto, bool)>(0, 4);
+			tuple.Item1.Should().Be(s_dto.TheString);
+			tuple.Item2.Should().BeEquivalentTo(new ItemDto { TheInt32 = s_dto.TheInt32, TheInt64 = s_dto.TheInt64 });
 			tuple.Item3.Should().BeTrue();
 
 			// tuple with two DTOs (needs NULL terminator)
-			Invoking(() => reader.Get<(ItemRecord, ItemRecord)>(0, 3)).Should().Throw<InvalidOperationException>();
+			Invoking(() => reader.Get<(ItemDto, ItemDto)>(0, 3)).Should().Throw<InvalidOperationException>();
 
 			// get nulls
 			reader.Read().Should().BeTrue();
 
 			// all nulls returns null DTO
-			reader.Get<ItemRecord>(0, 7).Should().BeNull();
+			reader.Get<ItemDto>(0, 7).Should().BeNull();
 		}
 
 		[Test]
@@ -268,15 +269,15 @@ namespace Faithlife.Data.Tests
 			reader.Read().Should().BeTrue();
 
 			// two DTOs
-			var tuple = reader.Get<(ItemRecord, ItemRecord)>(0, 5);
-			tuple.Item1.Should().BeEquivalentTo(new ItemRecord { TheString = s_record.TheString, TheInt64 = s_record.TheInt64 });
-			tuple.Item2.Should().BeEquivalentTo(new ItemRecord { TheBool = s_record.TheBool, TheDouble = s_record.TheDouble });
+			var tuple = reader.Get<(ItemDto, ItemDto)>(0, 5);
+			tuple.Item1.Should().BeEquivalentTo(new ItemDto { TheString = s_dto.TheString, TheInt64 = s_dto.TheInt64 });
+			tuple.Item2.Should().BeEquivalentTo(new ItemDto { TheBool = s_dto.TheBool, TheDouble = s_dto.TheDouble });
 
 			// get nulls
 			reader.Read().Should().BeTrue();
 
 			// two DTOs
-			tuple = reader.Get<(ItemRecord, ItemRecord)>(0, 5);
+			tuple = reader.Get<(ItemDto, ItemDto)>(0, 5);
 			tuple.Item1.Should().BeNull();
 			tuple.Item2.Should().BeNull();
 		}
@@ -293,18 +294,66 @@ namespace Faithlife.Data.Tests
 			reader.Read().Should().BeTrue();
 
 			// two DTOs
-			var tuple = reader.Get<(ItemRecord, ItemRecord)>(0, 2);
-			tuple.Item1.Should().BeEquivalentTo(new ItemRecord { TheString = s_record.TheString });
-			tuple.Item2.Should().BeEquivalentTo(new ItemRecord { TheInt64 = s_record.TheInt64 });
+			var tuple = reader.Get<(ItemDto, ItemDto)>(0, 2);
+			tuple.Item1.Should().BeEquivalentTo(new ItemDto { TheString = s_dto.TheString });
+			tuple.Item2.Should().BeEquivalentTo(new ItemDto { TheInt64 = s_dto.TheInt64 });
 
 			// get nulls
 			reader.Read().Should().BeTrue();
 
 			// two DTOs
-			tuple = reader.Get<(ItemRecord, ItemRecord)>(0, 2);
+			tuple = reader.Get<(ItemDto, ItemDto)>(0, 2);
 			tuple.Item1.Should().BeNull();
 			tuple.Item2.Should().BeNull();
 		}
+
+#if NET5_0
+		[Test]
+		[TestCase(typeof(ItemRecord), true)]
+		[TestCase(typeof(ItemDto), false)]
+		[TestCase(typeof(NonPositionalRecord), false)]
+		[TestCase(typeof(DtoWithConstructors), false)]
+		public void IsPositionalRecordTests(Type type, bool isPositionalRecord)
+		{
+			Assert.AreEqual(isPositionalRecord, DbValueTypeInfo.IsPositionalRecord(type));
+		}
+
+		[Test]
+		public void RecordTests()
+		{
+			using var connection = GetOpenConnection();
+			using var command = connection.CreateCommand();
+			command.CommandText = "select TheString, TheInt32, TheInt64, TheBool, TheSingle, TheDouble, TheBlob from items;";
+			using var reader = command.ExecuteReader();
+
+			// get non-nulls
+			reader.Read().Should().BeTrue();
+
+			// compare by members because the TheBlob is compared by reference in .Equals()
+			Func<EquivalencyAssertionOptions<ItemRecord>, EquivalencyAssertionOptions<ItemRecord>> configureEquivalency = options => options.ComparingByMembers<ItemRecord>();
+
+			// record
+			reader.Get<ItemRecord>(0, 7).Should().BeEquivalentTo(s_record, configureEquivalency);
+			reader.Get<ItemRecord>(0, 1).Should().BeEquivalentTo(new ItemRecord(s_record.TheString, default, default, default, default, default, default), configureEquivalency);
+			reader.Get<ItemRecord>(0, 0).Should().BeNull();
+			reader.Get<ItemRecord>(7, 0).Should().BeNull();
+
+			// tuple with record
+			var tuple = reader.Get<(string, ItemRecord, bool)>(0, 4);
+			tuple.Item1.Should().Be(s_record.TheString);
+			tuple.Item2.Should().BeEquivalentTo(new ItemRecord(default, s_record.TheInt32, s_record.TheInt64, default, default, default, default), configureEquivalency);
+			tuple.Item3.Should().BeTrue();
+
+			// tuple with two records (needs NULL terminator)
+			Invoking(() => reader.Get<(ItemRecord, ItemRecord)>(0, 3)).Should().Throw<InvalidOperationException>();
+
+			// get nulls
+			reader.Read().Should().BeTrue();
+
+			// all nulls returns null record
+			reader.Get<ItemRecord>(0, 7).Should().BeNull();
+		}
+#endif
 
 		[Test]
 		public void CaseInsensitivePropertyName()
@@ -315,8 +364,8 @@ namespace Faithlife.Data.Tests
 			using var reader = command.ExecuteReader();
 
 			reader.Read().Should().BeTrue();
-			reader.Get<ItemRecord>(0, 2)
-				.Should().BeEquivalentTo(new ItemRecord { TheString = s_record.TheString, TheInt64 = s_record.TheInt64 });
+			reader.Get<ItemDto>(0, 2)
+				.Should().BeEquivalentTo(new ItemDto { TheString = s_dto.TheString, TheInt64 = s_dto.TheInt64 });
 		}
 
 		[Test]
@@ -328,8 +377,8 @@ namespace Faithlife.Data.Tests
 			using var reader = command.ExecuteReader();
 
 			reader.Read().Should().BeTrue();
-			reader.Get<ItemRecord>(0, 2)
-				.Should().BeEquivalentTo(new ItemRecord { TheString = s_record.TheString, TheInt64 = s_record.TheInt64 });
+			reader.Get<ItemDto>(0, 2)
+				.Should().BeEquivalentTo(new ItemDto { TheString = s_dto.TheString, TheInt64 = s_dto.TheInt64 });
 		}
 
 		[Test]
@@ -341,7 +390,7 @@ namespace Faithlife.Data.Tests
 			using var reader = command.ExecuteReader();
 
 			reader.Read().Should().BeTrue();
-			Invoking(() => reader.Get<ItemRecord>(0, 2)).Should().Throw<InvalidOperationException>();
+			Invoking(() => reader.Get<ItemDto>(0, 2)).Should().Throw<InvalidOperationException>();
 		}
 
 		[Test]
@@ -356,13 +405,13 @@ namespace Faithlife.Data.Tests
 			reader.Read().Should().BeTrue();
 
 			// dynamic
-			((string) reader.Get<dynamic>(0, 7).TheString).Should().Be(s_record.TheString);
-			((bool) ((dynamic) reader.Get<object>(0, 7)).TheBool).Should().Be(s_record.TheBool);
+			((string) reader.Get<dynamic>(0, 7).TheString).Should().Be(s_dto.TheString);
+			((bool) ((dynamic) reader.Get<object>(0, 7)).TheBool).Should().Be(s_dto.TheBool);
 
 			// tuple with dynamic
 			var tuple = reader.Get<(string, dynamic, bool)>(0, 4);
-			tuple.Item1.Should().Be(s_record.TheString);
-			((long) tuple.Item2.TheInt64).Should().Be(s_record.TheInt64);
+			tuple.Item1.Should().Be(s_dto.TheString);
+			((long) tuple.Item2.TheInt64).Should().Be(s_dto.TheInt64);
 			tuple.Item3.Should().BeTrue();
 
 			// tuple with two dynamics (needs NULL terminator)
@@ -388,10 +437,10 @@ namespace Faithlife.Data.Tests
 			reader.Read().Should().BeTrue();
 
 			// dictionary
-			((string) reader.Get<Dictionary<string, object?>>(0, 7)["TheString"]!).Should().Be(s_record.TheString);
-			((int) reader.Get<IDictionary<string, object?>>(0, 7)["TheInt32"]!).Should().Be(s_record.TheInt32);
-			((long) reader.Get<IReadOnlyDictionary<string, object?>>(0, 7)["TheInt64"]!).Should().Be(s_record.TheInt64);
-			((bool) reader.Get<IDictionary>(0, 7)["TheBool"]!).Should().Be(s_record.TheBool);
+			((string) reader.Get<Dictionary<string, object?>>(0, 7)["TheString"]!).Should().Be(s_dto.TheString);
+			((int) reader.Get<IDictionary<string, object?>>(0, 7)["TheInt32"]!).Should().Be(s_dto.TheInt32);
+			((long) reader.Get<IReadOnlyDictionary<string, object?>>(0, 7)["TheInt64"]!).Should().Be(s_dto.TheInt64);
+			((bool) reader.Get<IDictionary>(0, 7)["TheBool"]!).Should().Be(s_dto.TheBool);
 
 			// get nulls
 			reader.Read().Should().BeTrue();
@@ -412,20 +461,20 @@ namespace Faithlife.Data.Tests
 			reader.Read().Should().BeTrue();
 
 			// object/dynamic
-			reader.Get<object>(0).Should().Be(s_record.TheString);
-			((bool) reader.Get<dynamic>(3)).Should().Be(s_record.TheBool);
+			reader.Get<object>(0).Should().Be(s_dto.TheString);
+			((bool) reader.Get<dynamic>(3)).Should().Be(s_dto.TheBool);
 
 			// tuple with object
 			var tuple = reader.Get<(string, object, long)>(0, 3);
-			tuple.Item1.Should().Be(s_record.TheString);
-			tuple.Item2.Should().Be(s_record.TheInt32);
-			tuple.Item3.Should().Be(s_record.TheInt64);
+			tuple.Item1.Should().Be(s_dto.TheString);
+			tuple.Item2.Should().Be(s_dto.TheInt32);
+			tuple.Item3.Should().Be(s_dto.TheInt64);
 
 			// tuple with three objects (doesn't need NULL terminator when the field count matches exactly)
 			var tuple2 = reader.Get<(object, object, object)>(0, 3);
-			tuple2.Item1.Should().Be(s_record.TheString);
-			tuple2.Item2.Should().Be(s_record.TheInt32);
-			tuple2.Item3.Should().Be(s_record.TheInt64);
+			tuple2.Item1.Should().Be(s_dto.TheString);
+			tuple2.Item2.Should().Be(s_dto.TheInt32);
+			tuple2.Item3.Should().Be(s_dto.TheInt64);
 
 			// get nulls
 			reader.Read().Should().BeTrue();
@@ -444,7 +493,7 @@ namespace Faithlife.Data.Tests
 			using var reader = command.ExecuteReader();
 
 			reader.Read().Should().BeTrue();
-			reader.Get<ItemRecord>().Should().BeEquivalentTo(s_record);
+			reader.Get<ItemDto>().Should().BeEquivalentTo(s_dto);
 		}
 
 		[Test]
@@ -456,9 +505,9 @@ namespace Faithlife.Data.Tests
 			using var reader = command.ExecuteReader();
 
 			reader.Read().Should().BeTrue();
-			reader.Get<long>(2).Should().Be(s_record.TheInt64);
-			reader.Get<long>("TheInt64").Should().Be(s_record.TheInt64);
-			reader.Get<long>(^5).Should().Be(s_record.TheInt64);
+			reader.Get<long>(2).Should().Be(s_dto.TheInt64);
+			reader.Get<long>("TheInt64").Should().Be(s_dto.TheInt64);
+			reader.Get<long>(^5).Should().Be(s_dto.TheInt64);
 		}
 
 		[Test]
@@ -470,10 +519,10 @@ namespace Faithlife.Data.Tests
 			using var reader = command.ExecuteReader();
 
 			reader.Read().Should().BeTrue();
-			reader.Get<(long, bool)>(2, 2).Should().Be((s_record.TheInt64, s_record.TheBool));
-			reader.Get<(long, bool)>(2..4).Should().Be((s_record.TheInt64, s_record.TheBool));
-			reader.Get<(long, bool)>("TheInt64", 2).Should().Be((s_record.TheInt64, s_record.TheBool));
-			reader.Get<(long, bool)>("TheInt64", "TheBool").Should().Be((s_record.TheInt64, s_record.TheBool));
+			reader.Get<(long, bool)>(2, 2).Should().Be((s_dto.TheInt64, s_dto.TheBool));
+			reader.Get<(long, bool)>(2..4).Should().Be((s_dto.TheInt64, s_dto.TheBool));
+			reader.Get<(long, bool)>("TheInt64", 2).Should().Be((s_dto.TheInt64, s_dto.TheBool));
+			reader.Get<(long, bool)>("TheInt64", "TheBool").Should().Be((s_dto.TheInt64, s_dto.TheBool));
 		}
 
 		private static IDbConnection GetOpenConnection()
@@ -502,7 +551,7 @@ namespace Faithlife.Data.Tests
 			return connection;
 		}
 
-		private class ItemRecord
+		private class ItemDto
 		{
 			public string? TheString { get; set; }
 			public int TheInt32 { get; set; }
@@ -513,12 +562,37 @@ namespace Faithlife.Data.Tests
 			public byte[]? TheBlob { get; set; }
 		}
 
+#if NET5_0
+#pragma warning disable CA1801,SA1313
+		private record ItemRecord(string? TheString, int TheInt32, long TheInt64, bool TheBool, float TheSingle, double TheDouble, byte[]? TheBlob, int TheOptionalInt32 = 42);
+#pragma warning restore CA1801,SA1313
+
+		private record NonPositionalRecord
+		{
+			public string? TheString { get; set; }
+		}
+
+		private class DtoWithConstructors
+		{
+			public DtoWithConstructors()
+			{
+			}
+
+			public DtoWithConstructors(string theString)
+			{
+				TheString = theString;
+			}
+
+			public string? TheString { get; set; }
+		}
+#endif
+
 		private enum Answer
 		{
 			FortyTwo = 42,
 		}
 
-		private static readonly ItemRecord s_record = new ItemRecord
+		private static readonly ItemDto s_dto = new ItemDto
 		{
 			TheString = "hey",
 			TheInt32 = 42,
@@ -528,5 +602,16 @@ namespace Faithlife.Data.Tests
 			TheDouble = 3.1415,
 			TheBlob = new byte[] { 0x01, 0xFE },
 		};
+
+#if NET5_0
+		private static readonly ItemRecord s_record = new ItemRecord(
+			TheString: "hey",
+			TheInt32: 42,
+			TheInt64: 42,
+			TheBool: true,
+			TheSingle: 3.14f,
+			TheDouble: 3.1415,
+			TheBlob: new byte[] { 0x01, 0xFE });
+#endif
 	}
 }